Unified checkbox, radio, value inputs and single selects
