--- tracker/trackers/react/src/trackedContexts/TrackedInputContext/index.ts
@@ -1,5 +1,0 @@
-/*
- * Copyright 2022 Objectiv B.V.
- */
-
-export * from './TrackedInputContext';
